+++ .coveragerc
@@ -1512,7 +1512,9 @@ omit =
     homeassistant/components/zwave_me/sensor.py
     homeassistant/components/zwave_me/siren.py
     homeassistant/components/zwave_me/switch.py
-
+    homeassistant/components/persistent_notification/__init__.py
+    homeassistant/helpers/entity.py
+    homeassistant/helpers/entity_id.py
 [report]
 # Regexes for lines to exclude from consideration
 exclude_lines =
