+++ homeassistant/components/persistent_notification/__init__.py
@@ -11,7 +11,9 @@ from homeassistant.components import websocket_api
 from homeassistant.const import ATTR_FRIENDLY_NAME
 from homeassistant.core import Context, HomeAssistant, ServiceCall, callback
 from homeassistant.helpers import config_validation as cv
-from homeassistant.helpers.entity import async_generate_entity_id
+
+# quick and dirty hack to resolve circular import problem
+from homeassistant.helpers.entity_id import async_generate_entity_id
 from homeassistant.helpers.typing import ConfigType
 from homeassistant.loader import bind_hass
 from homeassistant.util import slugify
